+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 1.0.0 (November 18th, 2020)
+
+Initial implementation.  Includes jobs:
+
+* db_fuel/dbee/query
+* db_fuel/dbee/range
+
 # 0.0.1
 
 Shell